+++ cstar/tests/unit_tests/entrypoint/test_worker.py
@@ -22,18 +22,18 @@ from cstar.entrypoint.config import (
     get_job_config,
     get_service_config,
 )
+from cstar.entrypoint.runner import (
+    XBlueprintRunner,
+    XRunnerRequest,
+    XRunnerResult,
+    create_parser,
+)
 from cstar.entrypoint.utils import (
     ARG_DIRECTIVES_URI_LONG,
     ARG_LOGLEVEL_LONG,
     ARG_LOGLEVEL_SHORT,
     ARG_URI_LONG,
     ARG_URI_SHORT,
-)
-from cstar.entrypoint.xrunner import (
-    XBlueprintRunner,
-    XRunnerRequest,
-    XRunnerResult,
-    create_parser,
 )
 from cstar.execution.file_system import RomsFileSystemManager
 from cstar.execution.handler import ExecutionHandler, ExecutionStatus
