+++ cstar/tests/unit_tests/orchestration/test_hello_app.py
@@ -14,7 +14,7 @@ from cstar.entrypoint.config import (
     get_job_config,
     get_service_config,
 )
-from cstar.entrypoint.xrunner import XRunnerRequest
+from cstar.entrypoint.runner import XRunnerRequest
 from cstar.execution.handler import ExecutionStatus
 from cstar.orchestration.models import Application
 from cstar.orchestration.serialization import deserialize, serialize
